Create README.md
Welcome to "RFM Customer Segmentation & Cohort Analysis Project". This is the first project of the Capstone Project Series, which consists of 4 different project that contain different scenarios.

This is a project which we will learn what is RFM, and how to apply RFM Analysis and Customer Segmentation using K-Means Clustering. Also we will improve our Data Cleaning, Data Visualization and Exploratory Data Analysis capabilities. On the other hand we will create Cohort and Conduct Cohort Analysis.